requirements.txt
torch>=2.0.0
torchvision
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.24.0
einops
Pillow
tqdm
peft>=0.6.0
datasets
scipy
ftfy
safetensors
matplotlib